+++ NEWS.md
@@ -1,5 +1,9 @@
 # cargo-mutants changelog
 
+## Unreleased
+
+- New: Summarize the overall number of mutants generated, caught, missed, etc, at the end.
+
 ## 0.2.7
 
 Released 2022-07-11